requirements.txt
google-generativeai
networkx
matplotlib
pandas
numpy
pillow
openpyxl
streamlit
flake8
